MFD-438 - add servcies and entites for user history except blood sugar

+++ src/MyFoodDoc.Application/Entites/TrackedValus/UserWeight.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace MyFoodDoc.Application.Entites.TrackedValus
 {
-    public class UserWeight : AbstractTrackedValue<int>
+    public class UserWeight : AbstractTrackedValue<decimal>
     {
         public string UserId { get; set; }
     }
